agent-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
langgraph==0.2.28
langchain-openai==0.2.3
langchain-core==0.3.10
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.9.2
